Add Core Data to the octomod

Add the shared dataset for this research question to the central octomod structure. Whenever this is used, it will replace the prior or old data.

core() adds a new dataset to the octomod object. It checks to see if data is already present.

change_core() changes the old dataset from the octomod object to an updated dataset. It does not re-do the hypothesis testing.

core(octomod, core)

change_core(octomod, new_core, ...)

remove_core(octomod, ...)

Arguments

octomod
Object of class octomod

core
Data frame or tibble object

new_core
A modified or new data frame to be used

...
For extensibility

Value

octomod object with core data attached

Examples

library(magrittr)

octomod() %>%
  core(mtcars) %>%
  change_core(iris)

#> 
#> This `octomod` has the following components. 
#> 
#> Core Variables 
#> -------------- 
#> [1] "Sepal.Length" "Sepal.Width"  "Petal.Length" "Petal.Width"  "Species"     
#> 
#> Arms 
#> ---- 
#> NULL
#> 
#> Equipment 
#> ---- 
#> NULL