Sharpen the Hypothesis Testing with Statistical Metrics #' #' This function adds a new column to any arm that has been passed to #' equip(). It only applies to parsnip models currently, as defined in the #' approach from the respective arm. It provides a test statistic on model #' fit for the original core data. This new column is named metric. It is #' intended to expand from the tidied column to include information about #' the overall model. See details for currently supported tests. #' #' It currently supports: #' #' - logistic_reg = The function will return only the models that are #' generated by parsnip::logistic_reg(). It generates a simple c-statistic (or AUC from ROC curve). The default for this is that the binary event has levels c(0, 1), with the event of interest being 1 (e.g. the second event).

Sharpen the Hypothesis Testing with Statistical Metrics #' #' This function adds a new column to any arm that has been passed to #' equip(). It only applies to parsnip models currently, as defined in the #' approach from the respective arm. It provides a test statistic on model #' fit for the original core data. This new column is named metric. It is #' intended to expand from the tidied column to include information about #' the overall model. See details for currently supported tests. #' #' It currently supports: #' #' - logistic_reg = The function will return only the models that are #' generated by parsnip::logistic_reg(). It generates a simple c-statistic (or AUC from ROC curve). The default for this is that the binary event has levels c(0, 1), with the event of interest being 1 (e.g. the second event).

sharpen(octomod, which_arms = NULL, ...)

Arguments

octomod
Object of class octomod

which_arms
Vector of names of arms that should be run. Defaults to all arms that have not yet been equipped..

...
To pass additional parameters as needed

Value

An octomod object with arms outfitted with the test statistics.

Examples

library(magrittr)
library(parsnip)
df <- mtcars
df$am <- factor(df$am)
df$vs <- factor(df$vs)

om <-
  octomod() %>%
  core(df) %>%
  arm(
    title = "log_model",
    plan = am + vs ~ mpg + disp + hp,
    pattern = "parallel",
    approach = logistic_reg() %>% set_engine("glm"),
  ) %>%
  equip() %>%
  sharpen(which_arms = "log_model")

#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred

#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred

#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred

#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred

#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred

#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred

#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred

#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred